+++ contracts/RariFundManager.sol
@@ -225,6 +225,7 @@ contract RariFundManager is Initializable, Ownable {
         _interestFeesGeneratedAtLastFeeRateChange = data.interestFeesGeneratedAtLastFeeRateChange;
         _interestFeesClaimed = data.interestFeesClaimed;
         _interestFeeRate = RariFundManager(_authorizedFundManagerDataSource).getInterestFeeRate();
+        _withdrawalFeeRate = RariFundManager(_authorizedFundManagerDataSource).getWithdrawalFeeRate();
     }
 
     /**
@@ -586,7 +587,7 @@ contract RariFundManager is Initializable, Ownable {
     /**
      * @dev Emitted when funds have been withdrawn from RariFund.
      */
-    event Withdrawal(string indexed currencyCode, address indexed sender, address indexed payee, uint256 amount, uint256 amountUsd, uint256 rftBurned);
+    event Withdrawal(string indexed currencyCode, address indexed sender, address indexed payee, uint256 amount, uint256 amountUsd, uint256 rftBurned, uint256 withdrawalFeeRate);
 
     /**
      * @notice Deposits funds from `msg.sender` to the Rari Stable Pool in exchange for RFT minted to `to`.
@@ -665,22 +666,15 @@ contract RariFundManager is Initializable, Ownable {
     }
 
     /**
-     * @dev Internal function to withdraw funds from the Rari Stable Pool to `msg.sender` in exchange for RFT burned from `from`.
-     * You may only withdraw currencies held by the fund (see `getRawFundBalance(string currencyCode)`).
-     * Please note that you must approve RariFundManager to burn of the necessary amount of RFT.
-     * @param from The address from which RFT will be burned.
+     * @dev Internal function to withdraw funds from pools if necessary for `RariFundController` to hold at least `amount` of actual tokens.
+     * This function was separated from `_withdrawFrom` to avoid the stack going too deep.
      * @param currencyCode The currency code of the token to be withdrawn.
-     * @param amount The amount of tokens to be withdrawn.
+     * @param amount The minimum amount of tokens that must be held by `RariFundController` after withdrawing.
      */
-    function _withdrawFrom(address from, string memory currencyCode, uint256 amount, uint256[] memory pricesInUsd) internal fundEnabled cachePoolBalances {
-        // Input validation
-        address erc20Contract = _erc20Contracts[currencyCode];
-        require(erc20Contract != address(0), "Invalid currency code.");
-        require(amount > 0, "Withdrawal amount must be greater than 0.");
-
+    function withdrawFromPoolsIfNecessary(string memory currencyCode, uint256 amount) internal {
         // Check contract balance of token and withdraw from pools if necessary
-        IERC20 token = IERC20(erc20Contract);
-        uint256 contractBalance = token.balanceOf(_rariFundControllerContract);
+        address erc20Contract = _erc20Contracts[currencyCode];
+        uint256 contractBalance = IERC20(erc20Contract).balanceOf(_rariFundControllerContract);
 
         for (uint256 i = 0; i < _poolsByCurrency[currencyCode].length; i++) {
             if (contractBalance >= amount) break;
@@ -700,10 +694,33 @@ contract RariFundManager is Initializable, Ownable {
         }
 
         require(amount <= contractBalance, "Available balance not enough to cover amount even after withdrawing from pools.");
+    }
+
+    /**
+     * @dev Internal function to withdraw funds from the Rari Stable Pool to `msg.sender` in exchange for RFT burned from `from`.
+     * You may only withdraw currencies held by the fund (see `getRawFundBalance(string currencyCode)`).
+     * Please note that you must approve RariFundManager to burn of the necessary amount of RFT.
+     * @param from The address from which RFT will be burned.
+     * @param currencyCode The currency code of the token to be withdrawn.
+     * @param amount The amount of tokens to be withdrawn.
+     * @return The amount withdrawn after the fee.
+     */
+    function _withdrawFrom(address from, string memory currencyCode, uint256 amount, uint256[] memory pricesInUsd) internal fundEnabled cachePoolBalances returns (uint256) {
+        // Input validation
+        address erc20Contract = _erc20Contracts[currencyCode];
+        require(erc20Contract != address(0), "Invalid currency code.");
+        require(amount > 0, "Withdrawal amount must be greater than 0.");
+
+        // Withdraw from pools if necessary
+        withdrawFromPoolsIfNecessary(currencyCode, amount);
 
         // Manually cache raw fund balance
         bool cacheSetPreviously = _rawFundBalanceCache >= 0;
         if (!cacheSetPreviously) _rawFundBalanceCache = toInt256(getRawFundBalance(pricesInUsd));
+
+        // Calculate withdrawal fee and amount after fee
+        uint256 feeAmount = amount.mul(_withdrawalFeeRate).div(1e18);
+        uint256 amountAfterFee = amount.sub(feeAmount);
 
         // Get withdrawal amount in USD
         uint256 amountUsd = amount.mul(pricesInUsd[_currencyIndexes[currencyCode]]).div(10 ** _currencyDecimals[currencyCode]);
@@ -711,11 +728,13 @@ contract RariFundManager is Initializable, Ownable {
         // Calculate RFT to burn
         uint256 rftAmount = getRftBurnAmount(from, amountUsd);
 
-        // Update net deposits, burn RFT, transfer funds to msg.sender, and emit event
+        // Update net deposits, burn RFT, transfer funds to msg.sender, transfer fee to _withdrawalFeeMasterBeneficiary, and emit event
         _netDeposits = _netDeposits.sub(int256(amountUsd));
         rariFundToken.fundManagerBurnFrom(from, rftAmount); // The user must approve the burning of tokens beforehand
-        token.safeTransferFrom(_rariFundControllerContract, msg.sender, amount);
-        emit Withdrawal(currencyCode, from, msg.sender, amount, amountUsd, rftAmount);
+        IERC20 token = IERC20(erc20Contract);
+        token.safeTransferFrom(_rariFundControllerContract, msg.sender, amountAfterFee);
+        token.safeTransferFrom(_rariFundControllerContract, _withdrawalFeeMasterBeneficiary, feeAmount);
+        emit Withdrawal(currencyCode, from, msg.sender, amount, amountUsd, rftAmount, _withdrawalFeeRate);
 
         // Update _rawFundBalanceCache
         _rawFundBalanceCache = _rawFundBalanceCache.sub(int256(amountUsd));
@@ -726,6 +745,9 @@ contract RariFundManager is Initializable, Ownable {
 
         // Clear _rawFundBalanceCache
         if (!cacheSetPreviously) _rawFundBalanceCache = -1;
+
+        // Return amount after fee
+        return amountAfterFee;
     }
 
     /**
@@ -734,9 +756,10 @@ contract RariFundManager is Initializable, Ownable {
      * Please note that you must approve RariFundManager to burn of the necessary amount of RFT.
      * @param currencyCode The currency code of the token to be withdrawn.
      * @param amount The amount of tokens to be withdrawn.
+     * @return The amount withdrawn after the fee.
      */
-    function withdraw(string calldata currencyCode, uint256 amount) external {
-        _withdrawFrom(msg.sender, currencyCode, amount, rariFundPriceConsumer.getCurrencyPricesInUsd());
+    function withdraw(string calldata currencyCode, uint256 amount) external returns (uint256) {
+        return _withdrawFrom(msg.sender, currencyCode, amount, rariFundPriceConsumer.getCurrencyPricesInUsd());
     }
 
     /**
@@ -746,9 +769,9 @@ contract RariFundManager is Initializable, Ownable {
      * @param from The address from which RFT will be burned.
      * @param currencyCodes The currency codes of the tokens to be withdrawn.
      * @param amounts The amounts of the tokens to be withdrawn.
-     * @return Boolean indicating success.
+     * @return Array of amounts withdrawn after fees.
      */
-    function withdrawFrom(address from, string[] calldata currencyCodes, uint256[] calldata amounts) external onlyProxy cachePoolBalances {
+    function withdrawFrom(address from, string[] calldata currencyCodes, uint256[] calldata amounts) external onlyProxy cachePoolBalances returns (uint256[] memory) {
         // Input validation
         require(currencyCodes.length > 0 && currencyCodes.length == amounts.length, "Lengths of currency code and amount arrays must be greater than 0 and equal.");
         uint256[] memory pricesInUsd = rariFundPriceConsumer.getCurrencyPricesInUsd();
@@ -757,10 +780,14 @@ contract RariFundManager is Initializable, Ownable {
         _rawFundBalanceCache = toInt256(getRawFundBalance(pricesInUsd));
 
         // Make withdrawals
-        for (uint256 i = 0; i < currencyCodes.length; i++) _withdrawFrom(from, currencyCodes[i], amounts[i], pricesInUsd);
+        uint256[] memory amountsAfterFees = new uint256[](currencyCodes.length);
+        for (uint256 i = 0; i < currencyCodes.length; i++) amountsAfterFees[i] = _withdrawFrom(from, currencyCodes[i], amounts[i], pricesInUsd);
 
         // Reset _rawFundBalanceCache
         _rawFundBalanceCache = -1;
+
+        // Return amounts withdrawn after fees
+        return amountsAfterFees;
     }
 
     /**
@@ -865,11 +892,6 @@ contract RariFundManager is Initializable, Ownable {
     event InterestFeeDeposit(address beneficiary, uint256 amountUsd);
 
     /**
-     * @dev Emitted when fees on interest are withdrawn.
-     */
-    event InterestFeeWithdrawal(address beneficiary, uint256 amountUsd, string currencyCode, uint256 amount);
-
-    /**
      * @dev Internal function to deposit all accrued fees on interest back into the fund on behalf of the master beneficiary.
      * @return Integer indicating success (0), no fees to claim (1), or no RFT to mint (2).
      */
@@ -918,41 +940,47 @@ contract RariFundManager is Initializable, Ownable {
     }
 
     /**
-     * @notice Withdraws all accrued fees on interest to the master beneficiary.
-     * @param currencyCode The currency code of the interest fees to be claimed.
-     */
-    function withdrawFees(string calldata currencyCode) external fundEnabled onlyRebalancer {
-        // Input validation
-        require(_interestFeeMasterBeneficiary != address(0), "Master beneficiary cannot be the zero address.");
-        address erc20Contract = _erc20Contracts[currencyCode];
-        require(erc20Contract != address(0), "Invalid currency code.");
-
-        // Get currency prices
-        uint256[] memory pricesInUsd = rariFundPriceConsumer.getCurrencyPricesInUsd();
-
-        // Manually cache raw fund balance (no need to check if set previously because the function is external)
-        _rawFundBalanceCache = toInt256(getRawFundBalance(pricesInUsd));
-
-        // Get unclaimed interest fees, calculate withdrawal amount, and validate
-        uint256 amountUsd = getInterestFeesUnclaimed();
-        uint256 amount = amountUsd.mul(10 ** _currencyDecimals[currencyCode]).div(pricesInUsd[_currencyIndexes[currencyCode]]);
-        require(amount > 0, "No new fees are available to claim.");
-
-        // Update claimed interest fees, transfer tokens, and emit event
-        _interestFeesClaimed = _interestFeesClaimed.add(amountUsd);
-        IERC20(erc20Contract).safeTransferFrom(_rariFundControllerContract, _interestFeeMasterBeneficiary, amount);
-        emit InterestFeeWithdrawal(_interestFeeMasterBeneficiary, amountUsd, currencyCode, amount);
-
-        // Reset _rawFundBalanceCache
-        _rawFundBalanceCache = -1;
-    }
-
-    /**
      * @dev Converts an unsigned uint256 into a signed int256.
      * @param value The uint256 to convert.
      */
     function toInt256(uint256 value) internal pure returns (int256) {
         require(value < 2 ** 255, "SafeCast: value doesn't fit in an int256");
         return int256(value);
+    }
+
+    /**
+     * @dev The current withdrawal fee rate (scaled by 1e18).
+     */
+    uint256 private _withdrawalFeeRate;
+
+    /**
+     * @dev The master beneficiary of withdrawal fees; i.e., the recipient of all withdrawal fees.
+     */
+    address private _withdrawalFeeMasterBeneficiary;
+
+    /**
+     * @dev Returns the withdrawal fee rate (proportion of every withdrawal taken as a service fee scaled by 1e18).
+     */
+    function getWithdrawalFeeRate() public view returns (uint256) {
+        return _withdrawalFeeRate;
+    }
+
+    /**
+     * @dev Sets the withdrawal fee rate.
+     * @param rate The proportion of every withdrawal taken as a service fee (scaled by 1e18).
+     */
+    function setWithdrawalFeeRate(uint256 rate) external fundEnabled onlyOwner {
+        require(rate != _withdrawalFeeRate, "This is already the current withdrawal fee rate.");
+        require(rate <= 1e18, "The withdrawal fee rate cannot be greater than 100%.");
+        _withdrawalFeeRate = rate;
+    }
+
+    /**
+     * @dev Sets the master beneficiary of withdrawal fees.
+     * @param beneficiary The master beneficiary of withdrawal fees; i.e., the recipient of all withdrawal fees.
+     */
+    function setWithdrawalFeeMasterBeneficiary(address beneficiary) external fundEnabled onlyOwner {
+        require(beneficiary != address(0), "Master beneficiary cannot be the zero address.");
+        _withdrawalFeeMasterBeneficiary = beneficiary;
     }
 }